# apps/dashboard/app/api/v1/missions/[missionId]/document/route.ts
import { NextResponse } from "next/server";
import { dataClient } from "@tloz/data";
import { authenticateRequest } from "../../../../../../lib/api-auth";
import { authorizeMissionOperation } from "../../../../../../lib/tloz-api-authorization";

export async function PUT(request: Request, { params }: { params: Promise<{ missionId: string }> }) {
  const auth = await authenticateRequest(request as Parameters<typeof authenticateRequest>[0]);
  if (auth instanceof Response) return auth;

  const { missionId } = await params;
  if (!missionId || missionId.length > 128) {
    return NextResponse.json(
      { error: { code: "INVALID_REQUEST", message: "missionId inválido.", requestId: crypto.randomUUID() } },
      { status: 400 }
    );
  }

  let body: { markdown?: string };
  try {
    body = await request.json();
  } catch {
    return NextResponse.json(
      { error: { code: "INVALID_REQUEST", message: "Cuerpo de solicitud inválido.", requestId: crypto.randomUUID() } },
      { status: 400 }
    );
  }

  if (typeof body.markdown !== "string") {
    return NextResponse.json(
      { error: { code: "INVALID_REQUEST", message: "markdown (string) es requerido.", requestId: crypto.randomUUID() } },
      { status: 400 }
    );
  }
  if (body.markdown.length > 20000) {
    return NextResponse.json(
      { error: { code: "INVALID_REQUEST", message: "markdown no puede superar 20000 caracteres.", requestId: crypto.randomUUID() } },
      { status: 400 },
    );
  }

  try {
    const permission = await authorizeMissionOperation(auth.user, missionId);
    if (!permission.allowed) return permission.response;
    const detail = await dataClient.tloz.saveMissionDocument(missionId, body.markdown);
    return NextResponse.json({ data: detail });
  } catch {
    return NextResponse.json(
      { error: { code: "INTERNAL_ERROR", message: "Error interno del servidor.", requestId: crypto.randomUUID() } },
      { status: 500 }
    );
  }
}