# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: TLOZ Data API
  version: 2.0.0-draft
  description: |
    Document-first API for TLOZ Projects, Missions and Inventory.
    The v1 routes remain available temporarily as a compatibility bridge.

    Authentication is via Bearer token (agent API key) or browser session cookie.
    If both are present, the Bearer token takes priority.
  contact:
    name: TLOZ
  license:
    name: Proprietary
    identifier: LicenseRef-Zivelo-Proprietary
servers:
  - url: /api/v1
    description: Legacy v1 compatibility API
security:
  - bearerToken: []
  - sessionCookie: []
...
paths:
...
  /projects/{projectId}:
    get:
      operationId: getProject
      summary: Get a TLOZ project by ID
      tags: [Projects]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project found.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      operationId: updateProject
      summary: Update a TLOZ project
      tags: [Projects]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                descriptionDetail:
                  type: string
                icon:
                  type: string
                color:
                  type: string
                  pattern: "^#[0-9a-fA-F]{6}$"
                status:
                  $ref: "#/components/schemas/ProjectStatus"
                type:
                  type: string
                ownerId:
                  type: string
                startDate:
                  type: string
                  format: date
                dueDate:
                  type: string
                  format: date
      responses:
        "200":
          description: Project updated.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Missions ───────────────────────────────────────────────────────
...
  /missions/query:
    post:
      operationId: queryMissions
      summary: Query missions by example
      description: Like GET /missions but filters are sent in the request body.
      tags: [Missions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectId:
                  type: string
                ownerId:
                  type: string
                status:
                  $ref: "#/components/schemas/MissionStatus"
                seasonId:
                  type: string
                episodeId:
                  type: string
                title:
                  type: string
                limit:
                  type: integer
                  minimum: 1
                  maximum: 100
                  default: 25
                cursor:
                  type: string
      responses:
        "200":
          description: Matching missions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissionListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /missions/{missionId}/document:
    put:
      operationId: saveMissionDocument
      summary: Save mission detail (Markdown with checklist)
      description: The Markdown detail is parsed for embedded checklist items.
      tags: [Missions]
      parameters:
        - $ref: "#/components/parameters/MissionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [markdown]
              properties:
                markdown:
                  type: string
                  maxLength: 20000
                  description: Full Markdown detail for the mission. Checklist items are parsed from `- [ ]` / `- [x]` syntax.
      responses:
        "200":
          description: Document saved. Returns updated mission detail.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/MissionDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Seasons ────────────────────────────────────────────────────────
...
  /quest-items/query:
    post:
      operationId: queryQuestItems
      summary: Query quest items by example
      description: Like GET /quest-items but filters are sent in the request body.
      tags: [Quest Items]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ownerId:
                  type: string
                status:
                  $ref: "#/components/schemas/InventoryStatus"
                category:
                  $ref: "#/components/schemas/InventoryCategory"
                limit:
                  type: integer
                  minimum: 1
                  maximum: 100
                  default: 25
                cursor:
                  type: string
      responses:
        "200":
          description: Matching quest items.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestItemListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /resources:
    get:
      operationId: listResources
      summary: List resources filtered by owner entity or type
      tags: [Resources]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: missionId
          in: query
          schema:
            type: string
        - name: projectId
          in: query
          schema:
            type: string
        - name: questItemId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/ResourceType"
      responses:
        "200":
          description: Matching resources.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /resources/query:
    post:
      operationId: queryResources
      summary: Query resources by example
      description: Like GET /resources but filters are sent in the request body.
      tags: [Resources]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                missionId:
                  type: string
                projectId:
                  type: string
                questItemId:
                  type: string
                type:
                  $ref: "#/components/schemas/ResourceType"
                limit:
                  type: integer
                  minimum: 1
                  maximum: 100
                  default: 25
                cursor:
                  type: string
      responses:
        "200":
          description: Matching resources.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Agents ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: Existing NextAuth browser session. Production may use the secure cookie prefix.
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: tloz_
      description: |
        New agent API keys use `tloz_` followed by 64 hex characters.
        Legacy `zaf_` keys remain valid during the v1 grace release.
        To authenticate, include the header:
        ```
        Authorization: Bearer tloz_<64-hex-chars>
        ```
        If both Bearer token and session cookie are present, the Bearer token takes priority.
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        minLength: 1
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    MissionId:
      name: missionId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  responses:
    BadRequest:
      description: Invalid path or query input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: An authenticated session or valid API key is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: The authenticated actor is not permitted to perform this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    InternalError:
      description: Unexpected server error. Details are recorded server-side.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    ApiError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum: [INVALID_REQUEST, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, INTERNAL_ERROR, DOCUMENT_INVALID, DOCUMENT_NOT_FOUND, DOCUMENT_REVISION_CONFLICT, PRECONDITION_REQUIRED, INVALID_ATTACHMENT_MANIFEST, ATTACHMENT_STORAGE_NOT_CONFIGURED, ATTACHMENT_STORAGE_ERROR, attachment_object_missing, attachment_object_invalid, batch_superseded]
            message:
              type: string
            fields:
              type: object
              description: Field-specific guidance for INVALID_REQUEST responses.
              additionalProperties:
                type: string
            requestId:
              type: string

    # ── Users ────────────────────────────────────────────────────────
...
    ProjectStatus:
      type: string
      enum: [planned, active, archived]
    Project:
      type: object
      additionalProperties: false
      required: [id, slug, name, description, color, icon, status, ownerId, startDate]
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        descriptionDetail:
          type: string
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        icon:
          type: string
        status:
          $ref: "#/components/schemas/ProjectStatus"
        ownerId:
          type: string
        startDate:
          type: string
          format: date
        dueDate:
          type: [string, "null"]
          format: date
...
    MissionStatus:
      type: string
      enum: [now, next, later, completed, blocked]
    MissionType:
      type: string
      enum: [main_quest, side_quest, farming_quest, exploration_quest]
    Mission:
      type: object
      additionalProperties: false
      required: [id, displayId, title, description, descriptionDetail, icon, type, status, ownerId, progress]
      properties:
        id:
          type: string
        displayId:
          type: string
        title:
          type: string
        description:
          type: string
          maxLength: 280
        descriptionDetail:
          type: string
          maxLength: 20000
        icon:
          type: string
        type:
          $ref: "#/components/schemas/MissionType"
        status:
          $ref: "#/components/schemas/MissionStatus"
        ownerId:
          type: string
        projectId:
          type: string
        seasonId:
          type: string
        episodeId:
          type: string
        dueDate:
          type: string
          format: date
        progress:
          type: integer
          minimum: 0
          maximum: 100
        completedAt:
          type: string
          format: date-time
        blockedReason:
          type: string
    MissionDetail:
      allOf:
        - $ref: "#/components/schemas/Mission"
        - type: object
          required: [checklist, checklistCount, completed, resources]
          properties:
            project:
              $ref: "#/components/schemas/Project"
            dependencies:
              type: array
              items:
                $ref: "#/components/schemas/Mission"
            questItems:
              type: array
              items:
                $ref: "#/components/schemas/QuestItem"
            checklist:
              type: array
              items:
                $ref: "#/components/schemas/ChecklistItem"
            checklistCount:
              type: integer
              minimum: 0
            completed:
              type: integer
              minimum: 0
            resources:
              type: array
              items:
                $ref: "#/components/schemas/Resource"
    MissionListResponse:
      type: object
      additionalProperties: false
      required: [data, nextCursor]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MissionDetail"
        nextCursor:
          type: [string, "null"]
...
    InventoryStatus:
      type: string
      enum: [locked, unlocked]
    InventoryCategory:
      type: string
      enum: [tool, access, asset, document, other]
    QuestItem:
      type: object
      additionalProperties: false
      required: [id, name, description, icon, status, category]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        status:
          $ref: "#/components/schemas/InventoryStatus"
        category:
          $ref: "#/components/schemas/InventoryCategory"
        ownerId:
          type: string
        acquiredAt:
          type: string
          format: date
    QuestItemListResponse:
      type: object
      additionalProperties: false
      required: [data, nextCursor]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/QuestItem"
        nextCursor:
          type: [string, "null"]

    # ── Resources ────────────────────────────────────────────────────
...
    ResourceType:
      type: string
      enum: [link, document, image, file, note]
...
    Resource:
      type: object
      additionalProperties: false
      required: [id, type, title]
      properties:
        id:
          type: string
        missionId:
          type: string
        projectId:
          type: string
        questItemId:
          type: string
        type:
          $ref: "#/components/schemas/ResourceType"
        icon:
          type: string
        title:
          type: string
        url:
          type: string
          format: uri
        fileId:
          type: string
        groupKey:
          type: string
        externalKey:
          type: string
        contentType:
          type: string
        sizeBytes:
          type: integer
        width:
          type: integer
        height:
          type: integer
        sourceRevision:
          type: string
    ResourceListResponse:
      type: object
      additionalProperties: false
      required: [data, nextCursor]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Resource"
        nextCursor:
          type: [string, "null"]

    # ── Container/Content v2 contract (routes land in TLO-0077) ─────
...
    ChecklistItem:
      type: object
      additionalProperties: false
      required: [id, missionId, title, completed, position]
      properties:
        id:
          type: string
        missionId:
          type: string
        title:
          type: string
        completed:
          type: boolean
        position:
          type: integer